Move client, server, redux counters to the separated features
