+++ packages/client/src/modules/counter/Counter.web.jsx
@@ -4,9 +4,9 @@ import Helmet from 'react-helmet';
 
 import { PageLayout } from '../common/components/web';
 import settings from '../../../../../settings';
-import ClientCounter from './clientCounter';
-import ServerCounter from './serverCounter';
-import ReduxCounter from './reduxCounter';
+import ClientCounter from './modules/clientCounter/containers/ClientCounter';
+import ServerCounter from './modules/serverCounter/containers/ServerCounter';
+import ReduxCounter from './modules/reduxCounter/containers/ReduxCounter';
 import translate from '../../i18n';
 
 const Counter = ({ t }) => (
