--- packages/client/src/modules/counter/clientCounter/index.js
@@ -1,3 +1,0 @@
-import ClientCounter from './containers/ClientCounter';
-
-export default ClientCounter;
