+++ packages/client/src/modules/counter/index.web.jsx
@@ -2,17 +2,11 @@ import React from 'react';
 import { Route } from 'react-router-dom';
 
 import Counter from './Counter';
-import resolvers from './clientCounter/resolvers';
-import reducers from './reduxCounter/reducers';
-import mergeLocalization from './helpers';
+import ClientCounter from './modules/clientCounter';
+import ReduxCounter from './modules/reduxCounter';
+import ServerCounter from './modules/serverCounter';
 import Feature from '../connector';
 
-export default new Feature({
-  route: <Route exact path="/" component={Counter} />,
-  resolver: resolvers,
-  reducer: { counter: reducers },
-  localization: {
-    ns: 'counter',
-    resources: mergeLocalization()
-  }
+export default new Feature(ClientCounter, ReduxCounter, ServerCounter, {
+  route: <Route exact path="/" component={Counter} />
 });
