+++ packages/client/src/modules/counter/modules/clientCounter/graphql/CounterQuery.client.graphql
