+++ packages/client/src/modules/counter/modules/clientCounter/index.native.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { createStackNavigator } from 'react-navigation';
+
+import translate from '../../../../i18n/index';
+import { HeaderTitle } from '../../../common/components/native/index';
+import ClientCounter from './containers/ClientCounter';
+import resolvers from './resolvers/index';
+import resources from './locales/index';
+import Feature from '../../../connector';
+
+const HeaderTitleWithI18n = translate('counter')(HeaderTitle);
+
+export default new Feature({
+  drawerItem: {
+    Counter: {
+      screen: createStackNavigator({
+        Counter: {
+          screen: ClientCounter
+        }
+      }),
+      navigationOptions: {
+        drawerLabel: <HeaderTitleWithI18n i18nKey="title" />
+      }
+    }
+  },
+  resolver: resolvers,
+  localization: { ns: 'counter', resources }
+});
