+++ packages/client/src/modules/counter/modules/clientCounter/index.web.jsx
@@ -1,0 +1,8 @@
+import resolvers from './resolvers/index';
+import resources from './locales/index';
+import Feature from '../../../connector';
+
+export default new Feature({
+  resolver: resolvers,
+  localization: { ns: 'counter', resources }
+});
