+++ packages/client/src/modules/counter/modules/clientCounter/locales/en/translations.json
