+++ packages/client/src/modules/counter/modules/clientCounter/locales/index.js
