+++ packages/client/src/modules/counter/modules/reduxCounter/index.web.jsx
@@ -1,0 +1,8 @@
+import resources from './locales/index';
+import Feature from '../../../connector';
+import reducers from './reducers/index';
+
+export default new Feature({
+  reducer: { counter: reducers },
+  localization: { ns: 'counter', resources }
+});
