+++ packages/client/src/modules/counter/modules/reduxCounter/locales/en/translations.json
