+++ packages/client/src/modules/counter/modules/reduxCounter/locales/index.js
