+++ packages/client/src/modules/counter/modules/serverCounter/components/ServerCounterView.native.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, Text, View } from 'react-native';
 
-import { Button, Loading, primary } from '../../../common/components/native';
+import { Button, Loading, primary } from '../../../../common/components/native/index';
 
 export const ServerCounterView = ({ t, children, counter, loading }) => {
   if (loading) {
