+++ packages/client/src/modules/counter/modules/serverCounter/containers/ServerCounter.spec.js
@@ -3,11 +3,11 @@ import chai from 'chai';
 import { step } from 'mocha-steps';
 import { render } from 'react-testing-library';
 
-import Renderer from '../../../../testHelpers/Renderer';
-import { click, find } from '../../../../testHelpers/testUtils';
+import Renderer from '../../../../../testHelpers/Renderer';
+import { click, find } from '../../../../../testHelpers/testUtils';
 import COUNTER_SUBSCRIPTION from '../graphql/CounterSubscription.graphql';
 import ServerCounter from './ServerCounter';
-import translate from '../../../../i18n';
+import translate from '../../../../../i18n/index';
 
 chai.should();
