+++ packages/client/src/modules/counter/modules/serverCounter/graphql/CounterQuery.graphql
