+++ packages/client/src/modules/counter/modules/serverCounter/graphql/CounterSubscription.graphql
