+++ packages/client/src/modules/counter/modules/serverCounter/index.native.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { createStackNavigator } from 'react-navigation';
+
+import translate from '../../../../i18n/index';
+import { HeaderTitle } from '../../../common/components/native/index';
+import ServerCounter from './containers/ServerCounter';
+import resources from './locales/index';
+import Feature from '../../../connector';
+
+const HeaderTitleWithI18n = translate('counter')(HeaderTitle);
+
+export default new Feature({
+  drawerItem: {
+    Counter: {
+      screen: createStackNavigator({
+        Counter: {
+          screen: ServerCounter
+        }
+      }),
+      navigationOptions: {
+        drawerLabel: <HeaderTitleWithI18n i18nKey="title" />
+      }
+    }
+  },
+  localization: { ns: 'counter', resources }
+});
