+++ packages/client/src/modules/counter/modules/serverCounter/index.web.jsx
@@ -1,0 +1,6 @@
+import resources from './locales/index';
+import Feature from '../../../connector';
+
+export default new Feature({
+  localization: { ns: 'counter', resources }
+});
