+++ packages/client/src/modules/counter/modules/serverCounter/locales/en/translations.json
