+++ packages/client/src/modules/counter/modules/serverCounter/locales/index.js
